prototype/requirements.txt
Django==1.5.1
argparse==1.2.1
django-social-auth==0.7.22
httplib2==0.8
oauth2==1.5.211
psycopg2==2.5
python-openid==2.2.5
wsgiref==0.1.2
